pyproject.toml
[project]
name = "rag-support-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "ragie (>=1.4.5,<2.0.0)",
    "google-genai (>=0.7.0,<0.8.0)",
    "pytelegrambotapi (>=4.26.0,<5.0.0)",
    "telegramify-markdown (>=0.4.1,<0.5.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "aiohttp (>=3.11.11,<4.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"

[tool.ruff]
extend-exclude = []
line-length = 88
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = [
  "E",     # pycodestyle
  "F",     # Pyflakes
  "B",     # flake8-bugbear
  "I",     # isort
  "W",     # pycodestyle
  "SIM",   # flake8-simplify
  "UP",    # pyupgrade
  "LOG",   # flake8-logging
  "G",     # flake8-logging-format
  "PT",    # flake8-pytest-style
  "S",     # flake8-bandit
  "DTZ",   # flake8-datetimez
  "Q",     # flake8-quotes
  "RET",   # flake8-return
  "TID",   # flake8-tidy-imports
  "PTH",   # flake8-use-pathlib
  "PERF",  # Perflint
  "FURB",  # refurb
  "RUF",   # Ruff-specific rules
  "C90",   # mccabe
  "N",     # pep8-naming
  "A",     # flake8-builtins
  "COM",   # flake8-commas
  "CPY",   # flake8-copyright
  "C4",    # flake8-comprehensions
  "ICN",   # flake8-import-conventions
  "T20",   # flake8-print
  "RSE",   # flake8-raise
  "TCH",   # flake8-type-checking
  "ARG",   # flake8-unused-arguments
  "FIX",   # flake8-fixme
  "ERA",   # eradicate
  "PL",    # Pylint
  "FLY",   # flynt
  # "D",     # pydocstyle
  "ANN",   # flake8-annotations
  "ASYNC", # flake8-async
  "T10",   # flake8-debugger
  "EM",    # flake8-errmsg
  "EXE",   # flake8-executable
  "FA",    # flake8-future-annotations
  "ISC",   # flake8-implicit-str-concat
  "INP",   # flake8-no-pep420
  "PIE",   # flake8-pie
  "PYI",   # flake8-pyi
  "SLF",   # flake8-self
  "SLOT",  # flake8-slots
  "INT",   # flake8-gettext
  "PGH",   # pygrep-hooks
  "DOC",   # pydoclint
]
ignore = []
exclude = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"